refactor(firestore): Refactoring tests
