+++ firestore/vector_search_basic.go
@@ -48,9 +48,9 @@ func vectorSearchBasic(w io.Writer, projectID string) error {
 		return err
 	}
 
-	// TODO(developer): Use vector query results.
-	_ = docs
-
+	for _, doc := range docs {
+		fmt.Fprintln(w, doc.Data()["name"])
+	}
 	return nil
 }
